// VIS_Application/app/Model/Masters/Configuration/EventCountDown.ts
export interface IEventCountDown
{
    Id: number;
    EventName: string;
    CountDownText: string;
    CountDownDate: string;
    CountDownTime: string;
    CountDownDateTime: string;
    NoOfDay: number;
    Active: boolean;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;
    CompanyId: number;
    EntityMessage: string;

}